requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow
python-decouple==3.8
openai==1.10.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
psycopg2-binary
